Add an IfStatement to check the return value of temp.delete().
This statement returns a value that is not checked.
The return value should be checked since it can indicate an unusual or unexpected function execution.
The statement returns false if the file could not be successfully deleted (rather than throwing an Exception).
If the result was not checked, developers would not notice if the statement signals an unexpected behavior by returning an atypical return value.
http://findbugs.sourceforge.net/bugDescriptions.html#RV_RETURN_VALUE_IGNORED_BAD_PRACTICE

+++ application/application-deployment-management/src/main/java/org/apache/aries/application/deployment/management/impl/DeploymentManifestManagerImpl.java
@@ -669,7 +669,9 @@ public class DeploymentManifestManagerImpl implements DeploymentManifestManager
       
       result.add(modelledResourceManager.getModelledResource(null, FileSystem.getFSRoot(temp)));
       // delete the temp file
-      temp.delete();
+      if(!temp.exists() && !temp.delete()) {
+      	throw new IOException("Can not delete the template file.");
+      }
       IOUtils.deleteRecursive(dir);
     }
     _logger.debug(LOG_EXIT, "getByValueBundles", new Object[]{result});
